% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizingFunctions.R
\name{createDemographicsTable}
\alias{createDemographicsTable}
\title{This function creates demographic data frame for database}
\usage{
createDemographicsTable(pathToResults, studyName)
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}

\item{studyName}{Name of the study}
}
\description{
This function creates demographic data frame for database
}
