% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executionFunctions.R
\name{createResultsDirectory}
\alias{createResultsDirectory}
\title{This function creates a "results" directory.}
\usage{
createResultsDirectory(db, pathToResults)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a "results" directory.
}
\keyword{internal}
