% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executionFunctions.R
\name{executeHeartFailureTrajectoryCostStudy}
\alias{executeHeartFailureTrajectoryCostStudy}
\title{This function creates a "results" directory.}
\usage{
executeHeartFailureTrajectoryCostStudy(
  dbms,
  connection,
  cdmSchema,
  cdmTmpSchema,
  cdmResultsSchema,
  studyName,
  pathToResults,
  databaseDescription,
  runTrajectoryCreation = TRUE
)
}
\arguments{
\item{dbms}{Database management system}

\item{connection}{DatabaseConnector object (connection)}

\item{cdmSchema}{Schema where ohdsi cdm tables are located}

\item{cdmTmpSchema}{Schema where the authenticated user can create temporary tables}

\item{cdmResultsSchema}{Schema where ohdsi result schemas are located}

\item{studyName}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}

\item{databaseDescription}{Description of the database}

\item{runTrajectoryCreation}{Boolean for running the first part of analysis}
}
\description{
This function creates a "results" directory.
}
