% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getFirstStateCostBarPlot}
\alias{getFirstStateCostBarPlot}
\title{This function creates the combined trajectory cost plots in relation with patients' starting state}
\usage{
getFirstStateCostBarPlot(databases, pathToResults)
}
\arguments{
\item{databases}{Names of the studies selected}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates the combined trajectory cost plots in relation with patients' starting state
}
\keyword{internal}
