% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getKMPlotPath}
\alias{getKMPlotPath}
\title{This function creates a path to Kaplan-Meier plot image}
\usage{
getKMPlotPath(db, from, to)
}
\arguments{
\item{db}{Database name}

\item{from}{Starting state id}

\item{to}{Ending date id}
}
\description{
This function creates a path to Kaplan-Meier plot image
}
\keyword{internal}
