% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getLRMatrixPlot}
\alias{getLRMatrixPlot}
\title{This function creates a heatmap plot from the LogRank test matrix}
\usage{
getLRMatrixPlot(db, pathToResults)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a heatmap plot from the LogRank test matrix
}
\keyword{internal}
