% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getMatrixDevPlot}
\alias{getMatrixDevPlot}
\title{This function creates a deviation heatmap plot from the transition matrix}
\usage{
getMatrixDevPlot(db, pathToResults, motherMatrix)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a deviation heatmap plot from the transition matrix
}
\keyword{internal}
