% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getMatrixPlot}
\alias{getMatrixPlot}
\title{This function creates a heatmap plot from the transition matrix#'}
\usage{
getMatrixPlot(db, pathToResults)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a heatmap plot from the transition matrix#'
}
\keyword{internal}
