% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getMotherMatrix}
\alias{getMotherMatrix}
\title{This function calculates the motherMatrix}
\usage{
getMotherMatrix(dbList, pathToResults)
}
\arguments{
\item{dbList}{List of selected study databases}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function calculates the motherMatrix
}
\keyword{internal}
