% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getStateCostBarPlot}
\alias{getStateCostBarPlot}
\title{This function creates the combined state cost plots}
\usage{
getStateCostBarPlot(databases, pathToResults)
}
\arguments{
\item{databases}{Names of the studies selected}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates the combined state cost plots
}
\keyword{internal}
