% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getSumMatrixPlot}
\alias{getSumMatrixPlot}
\title{This function creates a heatmap plot from the transition matrices}
\usage{
getSumMatrixPlot(motherMatrix)
}
\arguments{
\item{motherMatrix}{summarized matrix}
}
\description{
This function creates a heatmap plot from the transition matrices
}
\keyword{internal}
