% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getSunburstPlot}
\alias{getSunburstPlot}
\title{This function creates a sunburst plot of trajectories}
\usage{
getSunburstPlot(db, pathToResults)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a sunburst plot of trajectories
}
