% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{kmPlotGenerator}
\alias{kmPlotGenerator}
\title{This function creates a survival plot from the observed data vs generated data}
\usage{
kmPlotGenerator(observedData, generatedData, pathToResults, db)
}
\arguments{
\item{observedData}{The observed data}

\item{generatedData}{The generated data}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a survival plot from the observed data vs generated data
}
\keyword{internal}
