% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizingFunctions.R
\name{monetaryAnalysis}
\alias{monetaryAnalysis}
\title{This function evaluates cost of a mean patient trajectory for standard of care, telemonitoring and ICER, creates a summarizing table}
\usage{
monetaryAnalysis(pathToResults, studyName)
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}

\item{studyName}{Name of the study}
}
\description{
This function evaluates cost of a mean patient trajectory for standard of care, telemonitoring and ICER, creates a summarizing table
}
\keyword{internal}
