% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executionFunctions.R
\name{removeTempTables}
\alias{removeTempTables}
\title{This function creates a "results" directory.}
\usage{
removeTempTables(connection, dbms, cdmTmpSchema)
}
\arguments{
\item{connection}{DatabaseConnector object (connection)}

\item{dbms}{Database management system}

\item{cdmTmpSchema}{Schema where the authenticated user can create temporary tables}
}
\description{
This function creates a "results" directory.
}
\keyword{internal}
