% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executionFunctions.R
\name{standardizeSunburstPlot}
\alias{standardizeSunburstPlot}
\title{This function standardizes the sunburstplot, conf for same colors}
\usage{
standardizeSunburstPlot(db, pathToResults)
}
\arguments{
\item{db}{Name of the study}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function standardizes the sunburstplot, conf for same colors
}
\keyword{internal}
