% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizationFunctions.R
\name{getSumMatrixPlot}
\alias{getSumMatrixPlot}
\title{This function creates a heatmap plot from the transition matrices}
\usage{
getSumMatrixPlot(dbList, pathToResults)
}
\arguments{
\item{dbList}{List of selected study databases}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function creates a heatmap plot from the transition matrices
}
\keyword{internal}
